Initial 3DS support
UDP protocol changed to accommodate the extra buttons on a New 3DS.
No bells or whistles for the 3DS yet:
Extra buttons work, but no analog support for the cstick, touch screen
doesn't currently do anything. DS GUI is disabled on the 3DS for now.
Windows GUI is not updated, nor tested with 3DS.